requirements/requirements.txt
cryptography==2.8
Django==2.2.7
djangorestframework==3.10.3
drf-yasg==1.17.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-magic==0.4.15
PyYAML==5.1.2
requests==2.22.0
